Fix setting and updating size of transient's frame
Transient may now set `window-size-fixed' to `width', which messes with
`fit-frame-to-buffer'.  See https://github.com/minad/vertico/issues/532
(sic).

`transient--fit-window-to-buffer' binds `window-size-fixed' to `nil',
but this package wants/needs to use its own code to the sizing, so we
have to set the buffer-local value using the `:initialize' argument.

Transient resizes the transient window when updating it, because it may
start to show more bindings needing more space (try the "C-x" prefix).
That has to be adviced to use `fit-frame-to-buffer-1' instead of
`fit-window-to-buffer'.